+++ CHANGELOG.md
@@ -12,5 +12,6 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - `ruTorrent.edit_torrents()` method for editing trackers, private flag, and comments.
+- Command `xirvik rtorrent install-services` to install services for launchd and systemd.
 
 [unreleased]: https://github.com/Tatsh/ultra-todo/-/compare/v0.5.1...HEAD